Correct error in lower_case_table_names conditional
Summary:
WebScaleSQL Feature: Basic Bug Fixes

This fixes http://bugs.mysql.com/64347 reported in 5.1,
and apparently still not fixed in 5.6 or 5.7.

Use binary string comparisons in dboptions hash when
lower_case_table_names is zero, not when it is not zero.

Added a test case to expose this problem should it ever recur.

Test Plan: Included new test fails without this change, passes with it.

Reviewers: flamingcow, pivanof, jeremycole, andrew-ford, inaam-rana

Reviewed By: pivanof

CC: jtolmer, MarkCallaghan

Differential Revision: https://reviews.facebook.net/D15765